Move database components to their own package
