+++ features/auth/data/src/main/kotlin/com/nasdroid/auth/data/serverstore/database/AuthenticatedServerDatabase.kt
@@ -1,0 +1,9 @@
+package com.nasdroid.auth.data.serverstore.database
+
+import androidx.room.Database
+import androidx.room.RoomDatabase
+
+@Database(entities = [AuthenticatedServerDto::class], version = 2)
+internal abstract class AuthenticatedServerDatabase : RoomDatabase() {
+    abstract fun authenticatedServerDao(): AuthenticatedServerDao
+}
