+++ features/auth/data/src/main/kotlin/com/nasdroid/auth/data/serverstore/database/AuthenticatedServerDto.kt
@@ -1,0 +1,18 @@
+package com.nasdroid.auth.data.serverstore.database
+
+import androidx.room.ColumnInfo
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+
+@Entity(tableName = "authenticated_servers")
+internal data class AuthenticatedServerDto(
+    @PrimaryKey
+    @ColumnInfo("host_id")
+    val hostId: String,
+    @ColumnInfo(name = "server_address")
+    val serverAddress: String,
+    @ColumnInfo(name = "token")
+    val token: String,
+    @ColumnInfo("name")
+    val name: String,
+)
